feat: League Table
This PR adds functionality for showing the current league table. Right now, this table is static, but it should be possible during an active season for this table to update with live results as goals are scored, etc. For now, this is hard coded to the current season’s static endpoint. I suspect the table would update after any game is completed… but it might be as infrequent as the end of the day.

This also adds a navigation bar to move between the home page and league table. The current schedule will be added to this navigation once the season schedule is built.