feat: improve live game details on home page
Technically, dynamic caching rules cannot be applied on one endpoint, but there may be a way to have a control endpoint with no cache that can switch between returning calls from a hard cached endpoint, and a low cached endpoint… and when a result is done, it could also invalidate the cache on the static endpoint.

Unless Nextjs adds information about the cache duration in the response where this could be done all from one endpoint.